cloud_pipeline/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg[binary,pool]==3.2.9
paho-mqtt==2.1.0
